Python3: use int() part of a division
On Python3 the result of a division is not a decimal.
>>> 101/6
16.833333333333332

So the code "tlv / 16" to get the 4 MSB of a byte is bogus.
We now use "int(tlv / 16)".

+++ parseATR/parseATR.py
@@ -835,7 +835,7 @@ def compact_tlv(historical_bytes):
     if tlv is None:
         return text
 
-    tag = tlv / 16
+    tag = int(tlv / 16)
     len = tlv % 16
 
     text = list()
